implement synthesis improvements combining pr #6 strengths
- replace base64 with json.dumps() for cleaner text escaping
- convert to full async architecture with asyncio.create_subprocess_exec
- add configurable timeouts (5s eval, 10s text, 5s context)
- enhance error handling with detailed troubleshooting guidance
- split into modular structure: emacs_client.py and mcp_server.py
- fix narrowed buffer detection bug
- maintain backward compatibility with fallback functions
